Refactor: concept_id 검색 함수 별도 파일로 분리
